refactor: SchoolV1QueryService 통합 테스트 패키지 및 클래스명 변경
