--- backend/src/main/java/com/festago/school/dto/v1/SchoolFestivalArtistResponse.java
@@ -1,14 +1,0 @@
-package com.festago.school.dto.v1;
-
-import com.querydsl.core.annotations.QueryProjection;
-
-public record SchoolFestivalArtistResponse(
-    Long id,
-    String name,
-    String imageUrl
-) {
-
-    @QueryProjection
-    public SchoolFestivalArtistResponse {
-    }
-}
